requirements.txt
openai>=1.0.0
rich>=13.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
